requirements.txt
gradio
openai==0.28
requests
PILLOW
IPython